+++ R/msm-package.R
@@ -15,7 +15,7 @@
 #'
 #' @importFrom graphics plot persp contour image filled.contour legend lines par text
 #' @importFrom grDevices rainbow 
-#' @importFrom stats coef as.formula deriv dexp dnorm integrate logLik model.extract model.frame model.matrix na.fail na.omit na.pass numericDeriv optimHess pchisq pexp plogis pnorm qlogis qnorm quantile rbeta rbinom reformulate rexp rgamma rlnorm rnbinom rnorm rpois rt runif rweibull sd setNames terms uniroot
+#' @importFrom stats coef as.formula deriv dexp dnorm integrate logLik model.extract model.frame model.matrix na.fail na.omit na.pass numericDeriv optimHess pchisq pexp plogis pnorm qlogis qnorm quantile rbeta rbinom reformulate rexp rgamma rlnorm rnbinom rnorm rpois rt runif rweibull sd setNames terms uniroot reshape
 #' @importFrom utils head tail
 #' @importFrom mvtnorm rmvnorm
 #' @importFrom survival Surv survfit